ACTUALLY fix bigint

+++ db.sql
@@ -90,8 +90,8 @@ DROP TABLE IF EXISTS `user_servers`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `user_servers` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
-  `discord_user_id` int(11) DEFAULT NULL,
-  `discord_server_id` int(11) DEFAULT NULL,
+  `discord_user_id` bigint(20) DEFAULT NULL,
+  `discord_server_id` bigint(20) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -114,4 +114,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2022-02-12 22:47:28
+-- Dump completed on 2022-02-12 22:57:56
